Add --rm parameter to docker-compose run command examples
Each time you run a docker-compose run command it will create a new container and not reuse an older one. So when the command you've run is finished the container stop but remain in memory. In order to avoid creating deads containers each time you run a command it's better to add the --rm parameter.