Clarify name and source of URI for login
The URI used for login is not necessarily linked to the API entrypoint, it could even be on a different subdomain (in the case of a auth microservice for example). This change clarifies where the path comes from in the referenced documentation and changes the name in the code to make it clear that this is not necessarily linked tot he entrypoint.