Explain how the default IRI normalization works
The serialiser an (and will) turn referenced ApiResources into IRIs - if it knows the referenced item is an ApiResource. It determines this based on the Doctrine annotation, declared type or DocBlock type annotation. If an ApiResource is not a Doctrine entity (which is the recommended approach for non-prototyping apps), it is very easy to not "get" this default behaviour if no type is declared.