Update the environment variables used for Mercure
Since API Platform 2.6.5 the `mercure.hub_url` configuration is ignored if the Mercure bundle used is version ^0.3 and includes the `Discovery` class

The environment variables for the Mercure component are different so I have updated this documentation. I have not mentioned there is an option to use the `hub_url` configuration option. It may be worth mentioning in a comment on the configuration that the option is deprecated.

also see: https://github.com/api-platform/core/issues/4451

The link to the Symfony external parameter configuration was also broken.