Unprocessable test
If you consider that the reader come on this page following the steps from "The distribution: create powerful apis with ease", at this point the login question has not yet been considered, so testLogin() can't success. By the way a following paragraph says "If everything is working properly, you should see `OK (5 tests, 17 assertions)`." ; that's true if testLogin() is removed. Moreover a similar test on login can be found on Core->Testing Utilities.
I think testLogin() should be removed from this page.