docs: add new search filters syntax and the new make filter command (#2203)
* docs(doctrine/filters): document new custom filter command and syntax

* docs(laravel/filters): document new custom filter command and syntax

* docs(filters): promote QueryParameter usage over ApiFilter

The `ApiFilter` attribute is deprecated and scheduled for removal in API Platform 5.0. This commit updates the documentation to strongly recommend the modern `QueryParameter` syntax.

This new approach offers better flexibility and more explicit, per-operation filter configuration. The examples have been updated to reflect this best practice.

* docs(filters): add documentation for new search filters

Document the newly introduced search filters (`IriFilter`, `ExactFilter`, `PartialSearchFilter`, `FreeTextQueryFilter`, `OrFilter`) and clarify the deprecation of the `SearchFilter`. Encourage users to migrate to `QueryParameter` attributes for better configurability and explicitness.

* fix: lint

* docs(filters): enhance validation

Add notes on automatic validation benefits and reference `ParameterValidatorProvider` for Symfony and Laravel. Clarify the usage of `ParameterExtension` for handling empty values in Doctrine ORM.